+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Description of the upcoming release here.
 
 - [#4](https://github.com/FuelLabs/sway-standard-implementations/pull/4) Adds initial CI.
 - [#1](https://github.com/FuelLabs/sway-standard-implementations/pull/1) Adds an implementation of SRC-14: Owned Proxy.
+- [#8](https://github.com/FuelLabs/sway-standard-implementations/pull/8) Adds SRC-14 owned proxy usage scripts and docs.
 
 ### Changed